% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formbls.R
\name{formbls}
\alias{formbls}
\title{Get BLAST data}
\usage{
formbls(flin, flout = NULL)
}
\arguments{
\item{flin}{chr string of path for Fasta file}

\item{flout}{chr string of file to save for BLAST, output returned to console if \code{NULL}}
}
\description{
Get relevant BLAST data from output file
}
\examples{
\dontrun{
flin <- 'inst/bls_ex.txt'
formbls(flin)
}
}

