% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmtch.R
\name{getmtch}
\alias{getmtch}
\title{Get OTU matches}
\usage{
getmtch(lnsin)
}
\arguments{
\item{lnsin}{chr string of text to parse for an individual OTU}
}
\value{
A \code{\link[base]{data.frame}} of sequence info extracted from the OTU
}
\description{
Get sequence mathces for an individual OTU from BLAST file
}
\details{
Used within \code{\link{formbls}}
}

