% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formfst.R
\name{formfst}
\alias{formfst}
\title{Format fasta}
\usage{
formfst(flin, flout = NULL)
}
\arguments{
\item{flin}{chr string of path for Fasta file}

\item{flout}{chr string of file to save for BLAST, output returned to console if \code{NULL}}
}
\description{
Format fasta file for use with Blast
}
\examples{
flin <- 'inst/fst_ex.fasta'
formfst(flin)
}

